// src/openapi/models/BetStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BetMonitor for generating matches and Odds
 * The server for simulating a betting portal.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BetStatusDto = {
    Win: 'WIN',
    Loss: 'LOSS',
    Pending: 'PENDING',
    Void: 'VOID'
} as const;
export type BetStatusDto = typeof BetStatusDto[keyof typeof BetStatusDto];


export function BetStatusDtoFromJSON(json: any): BetStatusDto {
    return BetStatusDtoFromJSONTyped(json, false);
}

export function BetStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BetStatusDto {
    return json as BetStatusDto;
}

export function BetStatusDtoToJSON(value?: BetStatusDto | null): any {
    return value as any;
}